requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.4.0
SQLAlchemy==1.0.9
WTForms==2.0.2
Werkzeug==0.11.2
click==6.7
gunicorn==19.4.1
httplib2==0.9.2
itsdangerous==0.24
psycopg2==2.6.1
python-dotenv==0.6.2
pytz==2016.10
requests==2.8.1
six==1.10.0
twilio==5.6.0
wsgiref==0.1.2
